+++ tutorial/2-control/3-dependencies.cc
@@ -16,7 +16,7 @@ const auto dep_dc = package_d_action.add(package_c_action);
 int
 main() {
   const flecsi::run::dependencies_guard dg;
-  const flecsi::runtime run;
+  flecsi::runtime run;
   flecsi::flog::add_output_stream("clog", std::clog, true);
   return run.control<control>();
 } // main