+++ tutorial/3-execution/main.cc
@@ -6,7 +6,7 @@
 int
 main() {
   const flecsi::run::dependencies_guard dg;
-  const flecsi::runtime run;
+  flecsi::runtime run;
   flecsi::flog::add_output_stream("clog", std::clog, true);
   return run.control<control>();
 } // main