+++ tutorial/6-topology/ntree.cc
@@ -166,7 +166,7 @@ int
 main(int argc, char ** argv) {
   flecsi::getopt()(argc, argv);
   const flecsi::run::dependencies_guard dg;
-  const flecsi::runtime run;
+  flecsi::runtime run;
   flecsi::flog::add_output_stream("clog", std::clog, true);
   return run.control<sph::control>();
 }